requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow-macos
seaborn
pyyaml